backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# HTTP Client
requests==2.31.0

# Configuration
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==7.0.0
isort==5.13.2
